requirements.txt
fastapi
uvicorn
requests
cryptography
pycryptodome
